PHPStan: уровень 7 с зафиксированным долгом, и починка найденного (#309)
Уровень поднят с 5 до 7. Привести весь код разом нельзя: из 906 находок 633 —
требование описать типы значений массивов в сотнях докблоков. Остаток заморожен
в phpstan-baseline.neon: новый код проверяется на 7-м целиком, старый долг
посчитан и виден, а тест 293 держит бюджет — число может только уменьшаться.

138 находок оказались ложными: все про PDO::query()/prepare() с объявленным
`PDOStatement|false`, тогда как соединение открыто в режиме исключений.

Починено настоящее: 28 мест `date($fmt, strtotime($x))` переведены на
DateFormatter::format() — на неразобранной дате strtotime() отдаёт false, и в
strict_types это TypeError, ронявший карту сайта и RSS целиком, а во вьюхах
печаталось 1 января 1970; заодно ушло расхождение таймзон. Плюс FormController
(?q[]=1 ронял страницу заявок), WebPush (непроверенный отказ openssl), Uploader
и MediaResizer (нулевой размер холста), Block, Widget, TOTP, CalendarGrid.

+++ app/Core/TOTP.php
@@ -111,7 +111,7 @@ final class TOTP
             if (strlen($chunk) < 5) {
                 $chunk = str_pad($chunk, 5, '0');
             }
-            $output .= self::BASE32_ALPHABET[bindec($chunk)];
+            $output .= self::BASE32_ALPHABET[(int) bindec($chunk)];
         }
 
         return $output;
